Removed the power menu
It is now done by wofi and I have set the keybinding for it in
hyprland.conf. I have created a shell script that does it and it works
fine. The shutdown option here was not even working because I needed to
use sudo for it to work. But I did that using wofi. So, I did not need
this anymore.